+++ src/env.cpp
@@ -32,6 +32,7 @@
 #include "path.h"
 #include "proc.h"
 #include "reader.h"
+#include "termsize.h"
 #include "wutil.h"  // IWYU pragma: keep
 
 /// Some configuration path environment variables.
@@ -46,8 +47,8 @@
 extern char **environ;
 
 /// The character used to delimit path and non-path variables in exporting and in string expansion.
-static const wchar_t PATH_ARRAY_SEP = L':';
-static const wchar_t NONPATH_ARRAY_SEP = L' ';
+static constexpr wchar_t PATH_ARRAY_SEP = L':';
+static constexpr wchar_t NONPATH_ARRAY_SEP = L' ';
 
 bool curses_initialized = false;
 
@@ -357,7 +358,13 @@ void env_init(const struct config_paths_t *paths /* or NULL */) {
     } else {
         vars.set_pwd_from_getcwd();
     }
-    vars.set_termsize();  // initialize the terminal size variables
+
+    // Initialize termsize variables.
+    auto termsize = termsize_container_t::shared().initialize(vars);
+    if (vars.get(L"COLUMNS").missing_or_empty())
+        vars.set_one(L"COLUMNS", ENV_GLOBAL, to_string(termsize.width));
+    if (vars.get(L"LINES").missing_or_empty())
+        vars.set_one(L"LINES", ENV_GLOBAL, to_string(termsize.height));
 
     // Set fish_bind_mode to "default".
     vars.set_one(FISH_BIND_MODE_VAR, ENV_GLOBAL, DEFAULT_BIND_MODE);
@@ -1214,17 +1221,6 @@ int env_stack_t::get_last_status() const { return acquire_impl()->perproc_data()
 
 void env_stack_t::set_last_statuses(statuses_t s) {
     acquire_impl()->perproc_data().statuses = std::move(s);
-}
-
-/// If they don't already exist initialize the `COLUMNS` and `LINES` env vars to reasonable
-/// defaults. They will be updated later by the `get_current_winsize()` function if they need to be
-/// adjusted.
-void env_stack_t::set_termsize() {
-    auto cols = get(L"COLUMNS");
-    if (cols.missing_or_empty()) set_one(L"COLUMNS", ENV_GLOBAL, DFLT_TERM_COL_STR);
-
-    auto rows = get(L"LINES");
-    if (rows.missing_or_empty()) set_one(L"LINES", ENV_GLOBAL, DFLT_TERM_ROW_STR);
 }
 
 /// Update the PWD variable directory from the result of getcwd().